Don't display non-visible highlight-adjacent nodes
Visibility Filters currently do not apply to upstream/downstream
adjacent nodes of highlighted nodes. For example, if Private methods are
de-selected in the Visibility Filters, Private methods that call or are
called by a highlighted node will be visible in the graph despite the
"Private functions" checkbox being toggled to off.

This change checks the Visibility Filters before painting
upstream/downstream nodes of highlighted nodes.